Strip the `l5d-dst-canonical` header from inbound requests
The `l5d-dst-canonical` header is used convey context from a client
proxy to server proxy.

There's no use in sending this to applications, though; and, in some
cases it can surprise applications (e.g. that whitelist allowed
headers).

This change strip these headers as they are received on the inbound
proxy (after they are used to discover the logical dst).